Remove formatting logic from RowingStatistics
Remove logic of formatting metrics in RowingStatistics and move these to
the consuming end (e.g. client).
